fix: added average data calculation after edit/delete [ref:#1894]
