+++ apps/gauzy/src/app/@shared/expenses/expenses-mutation/expenses-mutation.component.ts
@@ -37,8 +37,6 @@ import { TranslationBaseComponent } from '../../language-base/translation-base.c
 import { ErrorHandlingService } from '../../../@core/services/error-handling.service';
 import { IOrganizationExpenseCategory } from '../../../../../../../libs/models/src/lib/organization-expense-category.model';
 import { OrganizationExpenseCategoriesService } from '../../../@core/services/organization-expense-categories.service';
-import { EmployeesService } from '../../../@core/services';
-import { EmployeeStatisticsService } from '../../../@core/services/employee-statistics.service';
 
 @Component({
 	selector: 'ga-expenses-mutation',
@@ -91,13 +89,11 @@ export class ExpensesMutationComponent extends TranslationBaseComponent
 		private organizationVendorsService: OrganizationVendorsService,
 		private store: Store,
 		private readonly organizationContactService: OrganizationContactService,
-		private employeesService: EmployeesService,
 		private readonly organizationProjectsService: OrganizationProjectsService,
 		private readonly expenseCategoriesStore: OrganizationExpenseCategoriesService,
 		private readonly toastrService: NbToastrService,
 		readonly translateService: TranslateService,
-		private errorHandler: ErrorHandlingService,
-		private employeeStatisticsService: EmployeeStatisticsService
+		private errorHandler: ErrorHandlingService
 	) {
 		super(translateService);
 	}
@@ -171,31 +167,7 @@ export class ExpensesMutationComponent extends TranslationBaseComponent
 				this.form.value
 			)
 		);
-		await this.getEmployeeStatistics(
-			this.employeeSelector.selectedEmployee.id
-		);
-		this.employeesService.update(
-			this.employeeSelector.selectedEmployee.id,
-			{
-				averageExpenses: this.averageExpense
-			}
-		);
 	}
-
-	async getEmployeeStatistics(id) {
-		const statistics = await this.employeeStatisticsService.getStatisticsByEmployeeId(
-			id
-		);
-		this.averageExpense = this.countStatistic(statistics.expenseStatistics);
-	}
-
-	countStatistic(data: number[]) {
-		return data.filter(Number).reduce((a, b) => a + b, 0) !== 0
-			? data.filter(Number).reduce((a, b) => a + b, 0) /
-					data.filter(Number).length
-			: 0;
-	}
-
 	addNewCategory = async (
 		name: string
 	): Promise<IOrganizationExpenseCategory> => {
